refactor: migrate setuptools to pyproject
