+++ pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "snmp-server"
+version = "0.1.6"
+description = "SNMP server as a pytest plugin"
+authors = [
+    {name = "Dmitry Alimov",email = "dvalimov@gmail.com"}
+]
+license = {text = "MIT"}
+readme = "README.rst"
+requires-python = ">=3.8"
+packages = [
+    { include = "pytest_snmpserver" },
+]
+
+dependencies = [
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.3.5"
+
